"job" execution mock up stuff
xjob.c is meant to contain all "state machines" that handle long running drive
operations, like seeking, data transfers, batch jobs, etc.

core1 is reserved for this purpose. It feels slightly wrong because we're using
a powerful CPU core to busy-wait for conditions to change 99% of the time.
Alternatives considered:

 - Explicit yields; write the code so it never blocks; instead of sleep_us(5),
   you return from the function, setup a future timestamp for when the next
   step needs to be executed. Rejcted because: hard to read and write

 - Coroutines / co-operative multitasking using setjmp/longjmp. Rejected
   because: is this possible, or a hack only possible on some platforms? I
   don't know!

 - "Byte code" programs? :-) Rejected because: complicated! (might be possible
   to write using C-macros, but...)