WIP on a full loopback test
the idea is to use BIT0/1 to transmit some "known data" back to READ_DATA/CLOCK
at 10MHz. this replaces/removes DEBUGCLK_10MHZ.

it's meant to be used as a self-check; to see if we can transmit a lot of
non-trivial data (see loopback_test_generate_data()) without losing data. it's
an especially valuable test for the differential line driver/receiver circuitry
needed to speak with the drive.