stuff
dedicated window + telemetry.log in spcfront. we saw some things working on a
real CDC9762 drive, but a lot of things didn't, and when you start the process
of elimination it's kinda bad to only be 99% sure what you saw :) so the
telemetry.log is meant to help with that. and the dedicated window just
contains the things we're likely to try, instead of everything.

also a warning when downloading files containing all zeroes (happened a lot
during real-life testing)

also a small bitdump tool for looking at nrz data

also fixing xop.c "pin wait" stuff: caller must decide whether or not to check
for errors during pin wait (specifically, checking for errors while doing unit
selection is bad, because the lack of UNIT SELECTED is also an error :)